+++ hydra-node/test/Hydra/Chain/Direct/Contract/Mutation.hs
@@ -146,6 +146,7 @@ import qualified Hydra.Chain.Direct.Fixture as Fixture
 import Hydra.Chain.Direct.Tx (assetNameFromVerificationKey)
 import qualified Hydra.Contract.Head as Head
 import qualified Hydra.Contract.HeadState as Head
+import Hydra.Data.ContestationPeriod
 import qualified Hydra.Data.Party as Data (Party)
 import Hydra.Ledger.Cardano (genKeyPair, genOutput, genVerificationKey, renderTxWithUTxO)
 import Hydra.Ledger.Cardano.Evaluate (evaluateTx)
@@ -644,12 +645,13 @@ replacePolicyIdWith originalPolicyId otherPolicyId output =
 
 replaceSnapshotNumber :: Head.SnapshotNumber -> Head.State -> Head.State
 replaceSnapshotNumber snapshotNumber = \case
-  Head.Closed{parties, utxoHash, contestationDeadline, headId, contesters} ->
+  Head.Closed{parties, utxoHash, contestationDeadline, headId, contesters, contestationPeriod} ->
     Head.Closed
       { Head.parties = parties
       , Head.snapshotNumber = snapshotNumber
       , Head.utxoHash = utxoHash
       , Head.contestationDeadline = contestationDeadline
+      , Head.contestationPeriod = contestationPeriod
       , Head.headId = headId
       , Head.contesters = contesters
       }
@@ -670,12 +672,13 @@ replaceParties parties = \case
       , Head.utxoHash = utxoHash
       , Head.headId = headId
       }
-  Head.Closed{snapshotNumber, utxoHash, contestationDeadline, headId, contesters} ->
+  Head.Closed{snapshotNumber, utxoHash, contestationDeadline, headId, contesters, contestationPeriod} ->
     Head.Closed
       { Head.parties = parties
       , Head.snapshotNumber = snapshotNumber
       , Head.utxoHash = utxoHash
       , Head.contestationDeadline = contestationDeadline
+      , Head.contestationPeriod = contestationPeriod
       , Head.headId = headId
       , Head.contesters = contesters
       }
@@ -690,12 +693,13 @@ replaceUtxoHash utxoHash = \case
       , Head.utxoHash = utxoHash
       , Head.headId = headId
       }
-  Head.Closed{parties, snapshotNumber, contestationDeadline, headId, contesters} ->
+  Head.Closed{parties, snapshotNumber, contestationDeadline, headId, contesters, contestationPeriod} ->
     Head.Closed
       { Head.parties = parties
       , Head.snapshotNumber = snapshotNumber
       , Head.utxoHash = utxoHash
       , Head.contestationDeadline = contestationDeadline
+      , Head.contestationPeriod = contestationPeriod
       , Head.headId = headId
       , Head.contesters = contesters
       }
@@ -703,12 +707,27 @@ replaceUtxoHash utxoHash = \case
 
 replaceContestationDeadline :: POSIXTime -> Head.State -> Head.State
 replaceContestationDeadline contestationDeadline = \case
-  Head.Closed{snapshotNumber, utxoHash, parties, headId, contesters} ->
+  Head.Closed{snapshotNumber, utxoHash, parties, headId, contesters, contestationPeriod} ->
     Head.Closed
       { snapshotNumber
       , utxoHash
       , parties
       , contestationDeadline
+      , contestationPeriod
+      , headId
+      , contesters
+      }
+  otherState -> otherState
+
+replaceContestationPeriod :: ContestationPeriod -> Head.State -> Head.State
+replaceContestationPeriod contestationPeriod = \case
+  Head.Closed{snapshotNumber, utxoHash, parties, headId, contesters, contestationDeadline} ->
+    Head.Closed
+      { snapshotNumber
+      , utxoHash
+      , parties
+      , contestationDeadline
+      , contestationPeriod
       , headId
       , contesters
       }
@@ -729,12 +748,13 @@ replaceHeadId headId = \case
       , Head.utxoHash = utxoHash
       , Head.headId = headId
       }
-  Head.Closed{snapshotNumber, utxoHash, contestationDeadline, parties, contesters} ->
+  Head.Closed{snapshotNumber, utxoHash, contestationDeadline, parties, contesters, contestationPeriod} ->
     Head.Closed
       { Head.parties = parties
       , Head.snapshotNumber = snapshotNumber
       , Head.utxoHash = utxoHash
       , Head.contestationDeadline = contestationDeadline
+      , Head.contestationPeriod = contestationPeriod
       , Head.headId = headId
       , Head.contesters = contesters
       }
@@ -742,12 +762,13 @@ replaceHeadId headId = \case
 
 replaceContesters :: [Plutus.PubKeyHash] -> Head.State -> Head.State
 replaceContesters contesters = \case
-  Head.Closed{snapshotNumber, utxoHash, contestationDeadline, parties, headId} ->
+  Head.Closed{snapshotNumber, utxoHash, contestationDeadline, parties, headId, contestationPeriod} ->
     Head.Closed
       { Head.parties = parties
       , Head.snapshotNumber = snapshotNumber
       , Head.utxoHash = utxoHash
       , Head.contestationDeadline = contestationDeadline
+      , Head.contestationPeriod = contestationPeriod
       , Head.headId = headId
       , Head.contesters = contesters
       }